apps/backend/requirements.txt
Django==6.0.2
djangorestframework==3.16.1
channels==4.3.2
daphne==4.2.1
channels-redis==4.3.0
celery==5.6.2
django-celery-results==2.6.0
psycopg[binary]==3.3.3
redis==7.2.0
django-cors-headers==4.9.0
python-dotenv==1.2.1
